feat: ts -> tsx로 확장자 변경 및 jsx문법 적용
